% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/romRDS.r
\name{get_dir_map}
\alias{get_dir_map}
\title{Returns list of mappings of file extentions to directories (used mainly by \code{rom_rds_read}). If options 'romRDS_dir_map' is set gets its values otherwise returns defauls from \code{get_dir_map_defaults()}}
\usage{
get_dir_map()
}
\value{
list with mappings
}
\description{
Returns list of mappings of file extentions to directories (used mainly by \code{rom_rds_read}). If options 'romRDS_dir_map' is set gets its values otherwise returns defauls from \code{get_dir_map_defaults()}
}
