% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/romRDS.r
\name{get_dir_map_defaults}
\alias{get_dir_map_defaults}
\title{Returns default options for mapping file extentions to directories (used mainly by \code{rom_rds_read})}
\usage{
get_dir_map_defaults()
}
\value{
list with mappings
}
\description{
Returns default options for mapping file extentions to directories (used mainly by \code{rom_rds_read})
}
