% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{get_path}
\alias{get_path}
\title{Makes path from file_name using directory to extention mappings specified in dir_map}
\usage{
get_path(
  file_names,
  dir_map = get_dir_map(),
  dir_only = FALSE,
  dir_terminate_with_sep = TRUE,
  make_dir = TRUE
)
}
\arguments{
\item{file_names}{File names. Could also be a paths but only file's base name will be used}

\item{dir_map}{A specification of mapping between file extention and directory. Can be a named list where names used as directory names and values if it is character are corresponding extentions that should go to the directory name or if it is list futher mappings of extentions to subdirectories in recursive manner}

\item{dir_only}{return dir path only}

\item{dir_terminate_with_sep}{if only dir path is returned whether to therminate it with dir separator (e.g. if it set then 'path/to/dor/' note separator '/' at the end)}

\item{make_dir}{Whether to create dir recursively if it does not exists}
}
\value{
A path as a string
}
\description{
Makes path from file_name using directory to extention mappings specified in dir_map
}
