% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/romRDS.r
\name{\%<-\%}
\alias{\%<-\%}
\title{A wrapper around romRDS}
\usage{
obj \%<-\% ...
}
\arguments{
\item{obj}{obj to assign to (also serves as .rds file name to make or search and load)}

\item{...}{what to assign (make)}
}
\description{
Note that ... should be enclosed in {} if it is not a single expression. For example if you do 'var \%<-\% 1 + 1' var would be 1 not 2 but 'var \%<-\% {1 + 1}' will work as expected
}
