% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/romRDS.r
\name{rom}
\alias{rom}
\title{Read or Make RDS}
\usage{
rom(...)
}
\arguments{
\item{...}{Parameters passed to 'romRDS'}
}
\value{
Nothing or name of the obj. Loads object into memory..
}
\description{
A shorthand for romRDS. Reads or makes .rds if file is not available
}
