% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/romRDS.r
\name{rom_rds}
\alias{rom_rds}
\title{Read or Make RDS}
\usage{
rom_rds(
  name,
  ...,
  rds_dir = "data/rds",
  rds_dir_ensure = TRUE,
  file_name_sufix = character(0),
  file_name_sufix_var = NULL,
  file_name_sufix_sep = ".",
  file_name_extention = "rds",
  name_has_extention = FALSE,
  rds_compress = FALSE,
  return_name = FALSE,
  do_not_make = FALSE,
  do_not_save = FALSE,
  quietly = FALSE,
  assign_to_name = FALSE
)
}
\arguments{
\item{name}{Name of file or object}

\item{...}{How to construct the object if it is not on disk}

\item{rds_dir}{Directory where to look for or save to .rds. Default is "data/rds"}

\item{rds_dir_ensure}{Make \code{rds_dir} direcory if it does not exist}

\item{file_name_sufix}{e.g., date or version.}

\item{file_name_sufix_var}{Either "date" or "time" for current date and time or NULL (default)}

\item{file_name_sufix_sep}{Default is "."}

\item{file_name_extention}{Default is ".rds"}

\item{name_has_extention}{Default is FALSE}

\item{rds_compress}{Default is FALSE. It is faster to read and write if rds is not compressed.}

\item{return_name}{Default is FALSE}

\item{do_not_make}{Default is FALSE}

\item{do_not_save}{Default is FALSE}

\item{quietly}{Default is FALSE}

\item{assign_to_name}{Default is FALSE}
}
\value{
Returns the object or name of the obj (if return_name is TRUE).
}
\description{
Reads or makes .rds if file is not available and loads the object into memory.
}
