% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rom_rds_read.r
\name{rom_rds_read}
\alias{rom_rds_read}
\title{Reads file file_name available from file_url and makes .rds for it for futher reuse}
\usage{
rom_rds_read(
  file_url,
  file_name = NULL,
  dir_map = getOption("romRDS_dir_map"),
  read_txt_as = "csv"
)
}
\arguments{
\item{file_url}{Files location. Can be a url link (should start with http\link{s}) or path to archive file (zip or rar)}

\item{file_name}{Name of the file. If omited the base name of \code{file_url} is used}

\item{dir_map}{Where to put the file. Mapping between destination directories and file's extention.}

\item{read_txt_as}{How to read txt files.}
}
\value{
path to rds file
}
\description{
Reads file file_name available from file_url and makes .rds for it for futher reuse
}
