% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{unzip_file}
\alias{unzip_file}
\title{Unzipts if it is zip file and return filenames}
\usage{
unzip_file(
  file_path,
  dir = "data",
  ext_as_sub_dir = TRUE,
  filter_files_regex = "^_|^\\\\."
)
}
\arguments{
\item{file_path}{file to unzip}

\item{dir}{output dir}

\item{ext_as_sub_dir}{use extention as subdirectory}

\item{filter_files_regex}{Filters filenames. Default is "^_|^\."}
}
\description{
Unzipts if it is zip file and return filenames
}
