% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rom_rds_from_file.r, R/rom_rds_read.r
\name{download_file}
\alias{download_file}
\title{Downloads file to a dir according to dir_map (see `` for details)}
\usage{
download_file(file_url, dir_map = getOption("romRDS_dir_map"))

download_file(file_url, dir_map = getOption("romRDS_dir_map"))
}
\arguments{
\item{file_url}{url}

\item{dir_map}{map ext to dir}
}
\value{
file name

file name
}
\description{
Downloads file to a dir according to dir_map (see `` for details)

Downloads file to a dir according to dir_map (see `` for details)
}
