% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/romRDS.r
\name{\%<-\%}
\alias{\%<-\%}
\title{A wrapper around romRDS}
\usage{
obj \%<-\% ...
}
\arguments{
\item{obj}{obj to assign to (also serves as .rds file name to make or search and load)}

\item{...}{what to assign (make)}
}
\description{
A wrapper around romRDS
}
