% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rom_rds_read.r
\name{rom_rds_read}
\alias{rom_rds_read}
\title{Reads file file_name available from file_url and makes .rds for it for futher reuse}
\usage{
rom_rds_read(
  file_url,
  file_name = NULL,
  dir_map = get_dir_map(),
  read_as_ext = c(csv = "txt"),
  file_url_globs_expantion = TRUE,
  copy_local_files = FALSE
)
}
\arguments{
\item{file_url}{Files location. Can be a url link (should start with http\link{s}) or path to archive file (zip or rar)}

\item{file_name}{Name of the file. If omited the base name of \code{file_url} is used}

\item{dir_map}{Where to put the file. Mapping between destination directories and file's extention.}

\item{read_as_ext}{Named character vertor that specifies how to read different file extentions. Names are supported methods ('csv', 'tsv', 'dta'), values are file expentions that should be processed with these methods.}

\item{file_url_globs_expantion}{Whether to expand (expect) globs in file_url. If so (default) then try expand url with 'Sys.glob'}

\item{copy_local_files}{}
}
\value{
path to rds file
}
\description{
Reads file file_name available from file_url and makes .rds for it for futher reuse
}
