% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/romRDS.r
\name{romRDS}
\alias{romRDS}
\title{Read or Make RDS}
\usage{
romRDS(obj.name, ..., dir = "rds", file.name.sufix = "",
  file.name.sufix.sep = ".", file.name.extention = ".rds",
  obj.name.has.extention = FALSE, compress.rds = FALSE,
  return.obj.name = FALSE, do.not.make = FALSE, quietly = FALSE)
}
\arguments{
\item{dir}{Directory where to look for or save to .rds. Default is "./rds"}

\item{file.name.sufix}{e.g., date or version. Default is ""}

\item{file.name.sufix.sep}{Default is "."}

\item{file.name.extention}{Default is ".rds"}

\item{obj.name.has.extention}{Default is FALSE}

\item{compress.rds}{Default is FALSE. It is faster to read and write if rds is not compressed.}

\item{return.obj.name}{Default is FALSE}

\item{do.not.make}{Default is FALSE}

\item{quietly}{Default is FALSE}
}
\value{
Nothing or name of the obj. Loads object into memory..
}
\description{
Reads or makes .rds if file is not available
}
